feat(temp): improve temperature control logic and optimize data types
-Change the data type of temperature configuration from float64 to int to simplify the data structure.
- Add output of temperature level and fan speed in main.go, and add log information.
- Refactored TempLevelCheck function to calculate temperature level and fan speed.
- Add maxInt function to get the maximum value in the temperature level array - Optimize the temperature monitoring logic to improve the stability and maintainability of the system

+++ model/level_config.go
@@ -10,26 +10,26 @@ type TempLevel struct {
 }
 
 type Level1FansSpeedConfig struct {
-	Temp  float64
+	Temp  int
 	Speed int
 }
 
 type Level2FansSpeedConfig struct {
-	Temp  float64
+	Temp  int
 	Speed int
 }
 
 type Level3FansSpeedConfig struct {
-	Temp  float64
+	Temp  int
 	Speed int
 }
 
 type Level4FansSpeedConfig struct {
-	Temp  float64
+	Temp  int
 	Speed int
 }
 
 type Level5FansSpeedConfig struct {
-	Temp  float64
+	Temp  int
 	Speed int
 }
